fix: resolve import.meta CJS warning in dev builds
Extract build-time string constants from src/lib/voice-runtime/assets.ts
into a new asset-constants.ts file. This removes the import.meta.env
usage from the transitive dependency of the Vite plugin config file
(vite.voice-runtime-assets.ts), which esbuild compiles as CJS.

When Vite processes config files, esbuild bundles them with CJS output
format. Including a file that references import.meta in that context
caused the '[empty-import-meta]' warning in both the frontend dev build
and the addin dev-linked build.

The fix:
- New file: asset-constants.ts contains only the pure string constants
- assets.ts imports from asset-constants.ts and re-exports them, keeping
  the public API unchanged; import.meta.env usage remains in this file
  which is only consumed in browser ESM builds
- vite.voice-runtime-assets.ts now imports from asset-constants.ts,
  avoiding any import.meta reference during Vite config processing

Fixes ERMAIN-312
